+++ hackerman/handlers/reverse_http.py
@@ -1,0 +1,5 @@
+from flask import flask, request
+from hackerman import utils
+import json
+
+class #CONT